# openapi/api.yaml
openapi: 3.0.0
info:
  title: Backend App API
  version: 1.0.0
  description: API for maintenance app backend
servers:
  - url: https://{ApiId}.execute-api.{Region}.amazonaws.com/v1
    description: Production server
paths:
...
  /admin/properties:
    post:
      summary: Create property (admin only)
      operationId: createProperty
      security:
        - bearerAuth: []
      description: Create a new property for a manager
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                owner_manager_sub:
                  type: string
                  description: Manager sub who will own this property
                name:
                  type: string
                  description: Property name
                address_line1:
                  type: string
                  description: Property address line 1
                address_line2:
                  type: string
                  nullable: true
                  description: Property address line 2
                suburb:
                  type: string
                  description: Property suburb
                state:
                  type: string
                  description: Property state
                postcode:
                  type: string
                  description: Property postcode
                country:
                  type: string
                  description: Property country
                timezone:
                  type: string
                  nullable: true
                  description: Property timezone
              required:
                - owner_manager_sub
                - name
                - address_line1
                - suburb
                - state
                - postcode
                - country
      responses:
        "200":
          description: Property created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: Property ID
                  owner_manager_sub:
                    type: string
                    description: Manager sub who owns this property
                  name:
                    type: string
                    description: Property name
                  address_line1:
                    type: string
                    description: Property address line 1
                  address_line2:
                    type: string
                    nullable: true
                    description: Property address line 2
                  suburb:
                    type: string
                    description: Property suburb
                  state:
                    type: string
                    description: Property state
                  postcode:
                    type: string
                    description: Property postcode
                  country:
                    type: string
                    description: Property country
                  timezone:
                    type: string
                    nullable: true
                    description: Property timezone
                  created_at:
                    type: string
                    format: date-time
                    description: Creation timestamp
                  updated_at:
                    type: string
                    format: date-time
                    description: Last update timestamp
                  archived_at:
                    type: string
                    format: date-time
                    nullable: true
                    description: Archive timestamp (null if not archived)
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      x-amazon-apigateway-integration:
        type: aws_proxy
        payloadFormatVersion: "2.0"
        httpMethod: POST
        uri:
          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${AdminCreatePropertyFunction.Arn}/invocations
  /admin/properties/{propertyId}:
    get:
      summary: Get property by ID (admin only)
      operationId: getProperty
      security:
        - bearerAuth: []
      description: Retrieve a specific property by ID
      parameters:
        - name: propertyId
          in: path
          required: true
          description: Property ID
          schema:
            type: string
      responses:
        "200":
          description: Property details
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: Property ID
                  ownerManagerSub:
                    type: string
                    description: Manager sub who owns this property
                  name:
                    type: string
                    description: Property name
                  addressLine1:
                    type: string
                    description: Property address line 1
                  addressLine2:
                    type: string
                    nullable: true
                    description: Property address line 2
                  suburb:
                    type: string
                    description: Property suburb
                  state:
                    type: string
                    description: Property state
                  postcode:
                    type: string
                    description: Property postcode
                  country:
                    type: string
                    description: Property country
                  timezone:
                    type: string
                    nullable: true
                    description: Property timezone
                  createdAt:
                    type: string
                    format: date-time
                    description: Creation timestamp
                  updatedAt:
                    type: string
                    format: date-time
                    description: Last update timestamp
                  archivedAt:
                    type: string
                    format: date-time
                    nullable: true
                    description: Archive timestamp (null if not archived)
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Property not found
      x-amazon-apigateway-integration:
        type: aws_proxy
        payloadFormatVersion: "2.0"
        httpMethod: POST
        uri:
          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${AdminGetPropertyFunction.Arn}/invocations

    patch:
      summary: Update property (admin only)
      operationId: updateProperty
      security:
        - bearerAuth: []
      description: Update an existing property
      parameters:
        - name: propertyId
          in: path
          required: true
          description: Property ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Property name
                address_line1:
                  type: string
                  description: Property address line 1
                address_line2:
                  type: string
                  nullable: true
                  description: Property address line 2
                suburb:
                  type: string
                  description: Property suburb
                state:
                  type: string
                  description: Property state
                postcode:
                  type: string
                  description: Property postcode
                country:
                  type: string
                  description: Property country
                timezone:
                  type: string
                  nullable: true
                  description: Property timezone
              required: []
      responses:
        "200":
          description: Property updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: Property ID
                  ownerManagerSub:
                    type: string
                    description: Manager sub who owns this property
                  name:
                    type: string
                    description: Property name
                  addressLine1:
                    type: string
                    description: Property address line 1
                  addressLine2:
                    type: string
                    nullable: true
                    description: Property address line 2
                  suburb:
                    type: string
                    description: Property suburb
                  state:
                    type: string
                    description: Property state
                  postcode:
                    type: string
                    description: Property postcode
                  country:
                    type: string
                    description: Property country
                  timezone:
                    type: string
                    nullable: true
                    description: Property timezone
                  createdAt:
                    type: string
                    format: date-time
                    description: Creation timestamp
                  updatedAt:
                    type: string
                    format: date-time
                    description: Last update timestamp
                  archivedAt:
                    type: string
                    format: date-time
                    nullable: true
                    description: Archive timestamp (null if not archived)
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Property not found
      x-amazon-apigateway-integration:
        type: aws_proxy
        payloadFormatVersion: "2.0"
        httpMethod: POST
        uri:
          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${AdminUpdatePropertyFunction.Arn}/invocations
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT